+++ R/digitizeRadii.R
@@ -9,6 +9,7 @@
 #' @param suffix See details in \code{\link{RFBCoptions}}.
 #' @param edgeIsAnnulus See details in \code{\link{RFBCoptions}}.
 #' @param windowSize See details in \code{\link{RFBCoptions}}.
+#' @param deviceType See details in \code{\link{RFBCoptions}}.
 #' @param popID See details in \code{\link{RFBCoptions}}.
 #' @param IDpattern See details in \code{\link{RFBCoptions}}.
 #' @param scaleBar See details in \code{\link{RFBCoptions}}.
@@ -64,7 +65,7 @@
 #' 
 digitizeRadii <- function(img,id,reading,suffix,
                           description,edgeIsAnnulus,popID,IDpattern,
-                          windowSize,
+                          windowSize,deviceType,
                           scaleBar,scaleBarLength,col.scaleBar,lwd.scaleBar,
                           scalingFactor,showTransect,snap2Transect,
                           col.transect,lwd.transect,
@@ -112,6 +113,7 @@ digitizeRadii <- function(img,id,reading,suffix,
   if (missing(windowSize)) windowSize <- iGetopt("windowSize")
   if (!is.numeric(windowSize)) STOP("'windowSize' must be numeric.")
   if (windowSize<=0) STOP("'windowSize' must be positive.")
+  if (missing(deviceType)) deviceType <- iGetopt("deviceType")
   if (missing(showInfo)) showInfo <- iGetopt("showInfo")
   if (missing(pos.info)) pos.info <- iGetopt("pos.info")
   if (missing(cex.info)) cex.info <- iGetopt("cex.info")
@@ -153,7 +155,7 @@ digitizeRadii <- function(img,id,reading,suffix,
     for (i in seq_along(img)) {
       digitizeRadii(img[i],id=id[i],reading,suffix,
                     description,edgeIsAnnulus,popID,IDpattern,
-                    windowSize,
+                    windowSize,deviceType,
                     scaleBar,scaleBarLength,col.scaleBar,lwd.scaleBar,
                     scalingFactor,showTransect,snap2Transect,
                     col.transect,lwd.transect,
@@ -166,7 +168,7 @@ digitizeRadii <- function(img,id,reading,suffix,
     ## Only one image to process
     dat <- iDigitizeRadii1(img,id,reading,suffix,
                            description,edgeIsAnnulus,popID,IDpattern,
-                           windowSize,
+                           windowSize,deviceType,
                            scaleBar,scaleBarLength,col.scaleBar,lwd.scaleBar,
                            scalingFactor,showTransect,snap2Transect,
                            col.transect,lwd.transect,
@@ -191,7 +193,7 @@ digitizeRadii <- function(img,id,reading,suffix,
 ########################################################################
 iDigitizeRadii1 <- function(img,id,reading,suffix,
                             description,edgeIsAnnulus,popID,IDpattern,
-                            windowSize,
+                            windowSize,deviceType,
                             scaleBar,scaleBarLength,col.scaleBar,lwd.scaleBar,
                             scalingFactor,showTransect,snap2Transect,
                             col.transect,lwd.transect,
@@ -206,7 +208,7 @@ iDigitizeRadii1 <- function(img,id,reading,suffix,
   msg2 <- "   'f'=finished, 'd'=delete, 'q'=abort, 'z'=restart"
 
   ## Loads image given in img ==================================================
-  windowInfo <- iGetImage(img,id,windowSize,
+  windowInfo <- iGetImage(img,id,windowSize,deviceType,
                           showInfo,pos.info,cex.info,col.info)
   DONE("Loaded the ",img," image.\n")
   
@@ -304,7 +306,7 @@ iDigitizeRadii1 <- function(img,id,reading,suffix,
     DONE("Processing is being RESTARTED as requested by the user.",
          " No file was written ",img,".\n")
     iDigitizeRadii1(img,id,reading,suffix,description,edgeIsAnnulus,popID,
-                    IDpattern,windowSize,scaleBar,
+                    IDpattern,windowSize,deviceType,scaleBar,
                     scaleBarLength,col.scaleBar,lwd.scaleBar,scalingFactor,
                     showTransect,snap2Transect,col.transect,lwd.transect,
                     pch.sel,col.sel,cex.sel,pch.del,col.del,showInfo,pos.info,